requirements.txt
affine==2.4.0
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
branca==0.8.1
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
contextily==1.6.2
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
distro==1.9.0
Flask==3.1.0
Flask-Cors==5.0.0
folium==0.19.4
fonttools==4.55.3
fpdf==1.7.2
geographiclib==2.0
geopandas==1.0.1
geopy==2.4.1
git-filter-repo==2.47.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mercantile==1.2.1
numpy==2.2.1
openai==1.59.5
packaging==24.2
pandas==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pyogrio==0.10.0
pyparsing==3.2.1
pypdfium2==4.30.1
pyproj==3.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
rasterio==1.4.3
reportlab==4.2.5
requests==2.32.3
shapely==2.0.6
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
Werkzeug==3.1.3
xyzservices==2024.9.0
flask-cors
Flask
Flask-Cors
psycopg2
requests
beautifulsoup4
openai
python-dotenv

